refactor(api): add dependency injection, project restruc
