+++ api/server.go
@@ -3,21 +3,22 @@ package api
 import (
 	"github.com/gin-gonic/gin"
 	db "github.com/reflection/frog_blossom_db/db/sqlc"
+	"github.com/reflection/frog_blossom_db/internal/handler"
 )
 
 // Server serves HTTP requets for CMS
 type Server struct {
-	store  *db.Store
+	Store  *db.Store
 	router *gin.Engine
 }
 
 // NewServer creates new HTTP server and sets up routing
 func NewServer(store *db.Store) *Server {
-	server := &Server{store: store}
+	server := &Server{Store: store}
 	router := gin.Default()
 
 	subrouter := router.Group("api/v1")
-	subrouter.POST("/users", server.createUsers)
+	subrouter.POST("/users", handler.CreateUsersHandler(store))
 
 	server.router = router
 	return server
@@ -25,8 +26,4 @@ func NewServer(store *db.Store) *Server {
 
 func (server *Server) Start(address string) error {
 	return server.router.Run(address)
-}
-
-func errorResponse(err error) gin.H {
-	return gin.H{"error": err.Error()}
 }
