--- main.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"database/sql"
-	"log"
-
-	_ "github.com/lib/pq"
-	"github.com/reflection/frog_blossom_db/api"
-	db "github.com/reflection/frog_blossom_db/db/sqlc"
-)
-
-const (
-	dbDriver      = "postgres"
-	dbSource      = "postgresql://root:secret@localhost:5432/frog-blossom-db?sslmode=disable"
-	serverAddress = "0.0.0.0:8080"
-)
-
-func main() {
-
-	conn, err := sql.Open(dbDriver, dbSource)
-	if err != nil {
-		log.Fatal("cannot connect to database:", err)
-	}
-
-	store := db.NewStore(conn)
-	server := api.NewServer(store)
-
-	err = server.Start(serverAddress)
-	if err != nil {
-		log.Fatal("cannot start server", err)
-	}
-}
